Separate integration tests from integration test harness
